Refactor participant mixins
Flat is better than nested.
