--- gratipay/models/participant/mixins/__init__.py
@@ -1,3 +1,0 @@
-from .identity import IdentityMixin as Identity
-
-__all__ = ['Identity']
